Remove deprecated \Serializable implementations
This concludes the migration started in https://github.com/nemiah/phpFinTS/pull/364.

This commit relatively mechanically removes the function implementations and `implements \Serializable` declarations from all classes that still contained them. From all the new-style `__unserialize()` functions, the fallbacks for recognizing the old format are removed and the unit test is updated accordingly. As a special case, `BaseDeg` and `BaseSegment` keep their `::serialize()` (not unserialize) member functions, because they're convenient short-hands used in a few places and possibly by client code for debugging/logging as well. The override in `AnonymousSegment` is moved from `__serialize()` to `serialize()` for simplicity.

This is breaking change because:
1. Data (meaning `FinTs` and `BaseMessage` instances) serialized before php-fints version 3.4 (or 3.3 if using PHP7.4+) cannot be unserialized anymore. This means that people can't upgrade from 3.3 to 4.0 straight away (or if they do, they have to throw away their serialized instances and log in from scratch). That said, 3.4 was released two and a half years ago; so users who haven't done any updates in that amount of time can be bothered to login once more.
2. The objects are no longer sub-classes of `\Serializable` and (with a few exceptions) their `->serialize()` and `->unserialize()` member functions are gone. That said, it's unlikely that any user of the library would have relied on these, as all example code and documentation has always pointed to the native top-level `\serialize()` and `\unserialize()` functions.